% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeze.R
\name{breeze}
\alias{breeze}
\title{Calculate}
\usage{
breeze(data = NULL, keep_character = FALSE)
}
\arguments{
\item{data}{xx}

\item{keep_character}{xx. Defaults to \code{FALSE}}
}
\value{
lorem ipsum
}
\description{
This function
}
\author{
Jeppe Vierø
}
