% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_breeze.R
\name{cross_breeze}
\alias{cross_breeze}
\title{Calculate}
\usage{
cross_breeze(data = NULL, cross_var = NULL, keep_character = FALSE)
}
\arguments{
\item{data}{xx}

\item{keep_character}{xx. Defaults to \code{FALSE}}
}
\value{
lorem ipsum
}
\description{
This function
}
\author{
Jeppe Vierø
}
