% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeze.R
\name{breeze}
\alias{breeze}
\title{Calculate dyadic distances between points \code{i} and \code{j} in a spatial dataset}
\usage{
breeze(data = NULL)
}
\arguments{
\item{data}{xx}
}
\value{
lorem ipsum
}
\description{
This function calculates the geodesic distance between any dyads (pairs of points \code{i} and \code{j}) in a spatial dataset and stores the result in a long tibble
}
\author{
Jeppe Vierø
}
